@@ -9,7 +9,7 @@ import { CommonDialogService } from '~common/dialog/services/common-dialog.servi
 import { ListPageDataService } from '~core/list-page/list-page-data.service';
 import { ListPageViewService } from '~core/list-page/list-page-view.service';
 import { SelectionWithFavoriteService } from '~core/list-page/selection-with-favorite.service';
-import { ListPageProviders } from '~core/list-page/list-page-providers.class';
+import { getProviders } from '~core/list-page/list-page-providers.class';
 import { TrackingComponent } from '~utils/tracking-component';
 
 /** since we use the sample component on different pages, this page will keep the methods clean */
