@@ -4,7 +4,7 @@ import { Router } from '@angular/router';
 import { InputDirective } from '~shared/inputs';
 import { FormControl } from '@angular/forms';
 import { TrackingComponent } from '~utils/tracking-component';
-import { ListPageProviders, ProviderKey } from '~core/list-page/list-page-providers.class';
+import { getProviders, ProviderKey } from '~core/list-page/list-page-providers.class';
 import { ListPageDataService } from '~core/list-page/list-page-data.service';
 import { ListPageViewService } from '~core/list-page/list-page-view.service';
 import { CommonDialogService } from '~common/dialog/services/common-dialog.service';
