@@ -7,14 +7,14 @@ import { ListPageDataService } from '~core/list-page/list-page-data.service';
 import { ListPageViewService } from '~core/list-page/list-page-view.service';
 import { SelectionWithFavoriteService } from '~core/list-page/selection-with-favorite.service';
 import { TrackingComponent } from '~utils/tracking-component';
-import { ListPageProviders } from '~core/list-page/list-page-providers.class';
+import { getProviders } from '~core/list-page/list-page-providers.class';
 
 @Component({
 	selector: 'products-page-app',
 	templateUrl: './products-page.component.html',
 	styleUrls: ['./products-page.component.scss'],
 	providers: [
-		ListPageProviders.getProviders('products-page', ERM.PRODUCT),
+		getProviders('products-page', ERM.PRODUCT),
 		CommonDialogService,
 		{ provide: ERM_TOKEN, useValue: ERM.PRODUCT }
 	]
