@@ -11,7 +11,7 @@ import { SelectionWithFavoriteService } from '~core/list-page/selection-with-fav
 import { ListPageDataService } from '~core/list-page/list-page-data.service';
 import { ListPageViewService } from '~core/list-page/list-page-view.service';
 import { TrackingComponent } from '~utils/tracking-component';
-import { ListPageProviders, ProviderKey } from '~core/list-page/list-page-providers.class';
+import { getProviders, ProviderKey } from '~core/list-page/list-page-providers.class';
 import { CommonDialogService } from '~common/dialog/services/common-dialog.service';
 
 @Component({
@@ -20,7 +20,7 @@ import { CommonDialogService } from '~common/dialog/services/common-dialog.servi
 	templateUrl: './project-products.component.html',
 	changeDetection: ChangeDetectionStrategy.OnPush,
 	providers: [
-		ListPageProviders.getProviders(ProviderKey.PROJECTS_PRODUCT, ERM.PRODUCT),
+		getProviders(ProviderKey.PROJECTS_PRODUCT, ERM.PRODUCT),
 		CommonDialogService,
 		{ provide: ERM_TOKEN, useValue: ERM.PRODUCT }]
 })
