@@ -25,7 +25,7 @@ import { CommonDialogService } from '~common/dialog/services/common-dialog.servi
 import { ListPageDataService } from '~core/list-page/list-page-data.service';
 import { ListPageViewService } from '~core/list-page/list-page-view.service';
 import { SelectionWithFavoriteService } from '~core/list-page/selection-with-favorite.service';
-import { ListPageProviders, ProviderKey } from '~core/list-page/list-page-providers.class';
+import { getProviders, ProviderKey } from '~core/list-page/list-page-providers.class';
 import { AutoUnsub } from '~utils/auto-unsub.component';
 
 @Component({
@@ -34,7 +34,7 @@ import { AutoUnsub } from '~utils/auto-unsub.component';
 	styleUrls: ['./project-workflow.component.scss'],
 	changeDetection: ChangeDetectionStrategy.OnPush,
 	providers: [
-		ListPageProviders.getProviders(ProviderKey.PROJECT_WORKFLOW, ERM.PRODUCT),
+		getProviders(ProviderKey.PROJECT_WORKFLOW, ERM.PRODUCT),
 		CommonDialogService,
 		{ provide: ERM_TOKEN, useValue: ERM.PRODUCT }
 	]
