@@ -3,7 +3,7 @@ import { ActivatedRoute, Router } from '@angular/router';
 import { CommonDialogService } from '~common/dialog/services/common-dialog.service';
 import { AbstractSampleCommonComponent } from '~common/sample/containers/abstract-sample-common.component';
 import { ListPageDataService } from '~core/list-page/list-page-data.service';
-import { ListPageProviders, ProviderKey } from '~core/list-page/list-page-providers.class';
+import { getProviders, ProviderKey } from '~core/list-page/list-page-providers.class';
 import { ListPageViewService } from '~core/list-page/list-page-view.service';
 import { SelectionWithFavoriteService } from '~core/list-page/selection-with-favorite.service';
 import { UserService } from '~entity-services';
@@ -18,7 +18,7 @@ import { FilterType } from '~shared/filters';
 	styleUrls: ['./supplier-samples.component.scss'],
 	changeDetection: ChangeDetectionStrategy.OnPush,
 	providers: [
-		ListPageProviders.getProviders(ProviderKey.SUPPLIER_SAMPLE, ERM.PRODUCT),
+		getProviders(ProviderKey.SUPPLIER_SAMPLE, ERM.PRODUCT),
 		CommonDialogService,
 		{ provide: ERM_TOKEN, useValue: ERM.PRODUCT }
 	]
