@@ -4,7 +4,7 @@ import { TaskService, UserService } from '~entity-services';
 import { ERM, Task } from '~models';
 import { CommonDialogService } from '~common/dialog/services/common-dialog.service';
 import { ListPageDataService } from '~core/list-page/list-page-data.service';
-import { ListPageProviders, ProviderKey } from '~core/list-page/list-page-providers.class';
+import { getProviders, ProviderKey } from '~core/list-page/list-page-providers.class';
 import { ListPageViewService } from '~core/list-page/list-page-view.service';
 import { SelectionWithFavoriteService } from '~core/list-page/selection-with-favorite.service';
 import { AbstractTaskCommonComponent } from '~common/task/containers/abstract-task-common.component';
@@ -14,7 +14,7 @@ import { AbstractTaskCommonComponent } from '~common/task/containers/abstract-ta
 	templateUrl: './my-tasks-page.component.html',
 	styleUrls: ['./my-tasks-page.component.scss'],
 	providers: [
-		ListPageProviders.getProviders(ProviderKey.MY_TASK, ERM.TASK),
+		getProviders(ProviderKey.MY_TASK, ERM.TASK),
 	]
 })
 export class MyTasksPageComponent extends AbstractTaskCommonComponent implements OnInit {
